+++ NEWS.md
@@ -3,6 +3,8 @@
 
 * `rim rm now supports removing multiple versions.
 
+* `rim system create-lib` can be now restricted to one or more R versions.
+
 # rim 0.1.0
 
 First pre-release.